chore: Renamed configuration module within analysis
