+++ ra2ce/analyses/analysis_config_data/analysis_ini_config_validator.py
@@ -20,7 +20,7 @@
 """
 
 
-from ra2ce.analyses.configuration.analysis_ini_config_data_validator_base import (
+from ra2ce.analyses.analysis_config_data.analysis_ini_config_data_validator_base import (
     IniConfigValidatorBase,
 )
 from ra2ce.common.validation.validation_report import ValidationReport
