+++ ra2ce/analyses/analysis_config_data/analysis_with_network_config.py
@@ -24,7 +24,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from ra2ce.analyses.configuration import AnalysisConfigBase, AnalysisConfigData
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase, AnalysisConfigData
 from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
