+++ ra2ce/analyses/analysis_config_data/readers/analysis_config_reader_factory.py
@@ -23,14 +23,14 @@
 from pathlib import Path
 from typing import Optional
 
-from ra2ce.analyses.configuration import AnalysisConfigBase
-from ra2ce.analyses.configuration.readers.analysis_config_reader_base import (
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase
+from ra2ce.analyses.analysis_config_data.readers.analysis_config_reader_base import (
     AnalysisConfigReaderBase,
 )
-from ra2ce.analyses.configuration.readers.analysis_with_network_config_reader import (
+from ra2ce.analyses.analysis_config_data.readers.analysis_with_network_config_reader import (
     AnalysisWithNetworkConfigReader,
 )
-from ra2ce.analyses.configuration.readers.analysis_without_network_config_reader import (
+from ra2ce.analyses.analysis_config_data.readers.analysis_without_network_config_reader import (
     AnalysisWithoutNetworkConfigReader,
 )
 from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
