--- ra2ce/analyses/configuration/__init__.py
@@ -1,4 +1,0 @@
-from ra2ce.analyses.configuration.analysis_config_base import (
-    AnalysisConfigBase,
-    AnalysisConfigData,
-)
