+++ ra2ce/ra2ce_handler.py
@@ -28,7 +28,7 @@ from typing import Optional
 
 from shapely.errors import ShapelyDeprecationWarning
 
-from ra2ce.analyses.configuration import AnalysisConfigBase
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase
 from ra2ce.configuration.config_factory import ConfigFactory
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
