+++ ra2ce/runners/direct_analysis_runner.py
@@ -22,7 +22,7 @@
 
 import logging
 
-from ra2ce.analyses.configuration import AnalysisConfigBase
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase
 from ra2ce.analyses.direct import analyses_direct
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.runners.analysis_runner_protocol import AnalysisRunner
