+++ ra2ce/runners/indirect_analysis_runner.py
@@ -20,7 +20,7 @@
 """
 
 
-from ra2ce.analyses.configuration import AnalysisConfigBase
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase
 from ra2ce.analyses.indirect import analyses_indirect
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.runners.analysis_runner_protocol import AnalysisRunner
