+++ tests/analyses/configuration/readers/test_analysis_without_network_config_reader.py
@@ -1,7 +1,7 @@
-from ra2ce.analyses.configuration.readers.analysis_config_reader_base import (
+from ra2ce.analyses.analysis_config_data.readers.analysis_config_reader_base import (
     AnalysisConfigReaderBase,
 )
-from ra2ce.analyses.configuration.readers.analysis_without_network_config_reader import (
+from ra2ce.analyses.analysis_config_data.readers.analysis_without_network_config_reader import (
     AnalysisWithoutNetworkConfigReader,
 )
