+++ tests/configuration/test_config_factory.py
@@ -1,4 +1,4 @@
-from ra2ce.analyses.configuration import AnalysisConfigBase
+from ra2ce.analyses.analysis_config_data import AnalysisConfigBase
 from ra2ce.configuration.config_factory import ConfigFactory
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
